requirements.txt
discord.py>=2.4.0
numpy>=2.0.0
pandas>=2.2.0
Pillow>=10.0.0
requests>=2.32.0
